feat: Add authorization error handling and new permission constants, enhance permission group management with additional checks, and improve user and permission service methods for better clarity and functionality.

+++ internal/permission/service.go
@@ -2,25 +2,20 @@ package permission
 
 import (
 	"context"
-	"errors"
+	"nh-be/constant"
+	"nh-be/utils"
+	"slices"
 	"time"
 
 	"github.com/google/uuid"
 )
 
-var (
-	ErrPermissionNotFound = errors.New("permission not found")
-	ErrPermissionGroupNotFound = errors.New("permission group not found")
-	ErrNotNullPermissions = errors.New("permissions can not be null")
-	ErrCannotDeleteSuperAdmin = errors.New("can not delete Super Admin. At least one must exist")
-	ErrRoleNameAlreadyExists = errors.New("role name already exists")
-)
-
-type Service interface {	
+type Service interface {
 	// Permission
 	GetAllPermissions(ctx context.Context, search string) ([]Permission, int64, error)
 	GetPermissionsByIDs(ctx context.Context, permissionIds []uuid.UUID) ([]Permission, error)
 	GetPermissionByID(ctx context.Context, id uuid.UUID) (*Permission, error)
+	GetUserPermissionCodeNames(ctx context.Context, userId uuid.UUID) ([]string, error)
 
 	// Permission Group
 	CreatePermissionGroup(ctx context.Context, permissionGroup *PermissionGroupInput) error
@@ -60,6 +55,20 @@ func (s *service) GetPermissionGroupsByIDs(ctx context.Context, permissionGroupI
 
 // Permission Implementations
 func (s *service) GetAllPermissions(ctx context.Context, search string) ([]Permission, int64, error) {
+	userId, err := utils.GetUserIdFromContext(ctx)
+	if err != nil {
+		return nil, 0, err
+	}
+
+	userPerm, err := s.GetUserPermissionCodeNames(ctx, userId)
+	if err != nil {
+		return nil, 0, err
+	}
+
+	if !slices.Contains(userPerm, constant.ViewPermissionGroup) && !slices.Contains(userPerm, constant.ManagePermissionGroup) {
+		return nil, 0, ErrForbidViewPermissionGroups
+	}
+
 	return s.permissionRepo.FindAllPermissions(ctx, search)
 }
 
@@ -67,10 +76,28 @@ func (s *service) GetPermissionByID(ctx context.Context, id uuid.UUID) (*Permiss
 	return s.permissionRepo.FindPermissionByID(ctx, id)
 }
 
+func (s *service) GetUserPermissionCodeNames(ctx context.Context, userId uuid.UUID) ([]string, error) {
+	return s.permissionRepo.FindUserPermissionCodeNames(ctx, userId)
+}
+
 // Permission Group Implementations
 func (s *service) CreatePermissionGroup(ctx context.Context, permissionGroup *PermissionGroupInput) error {
+	userId, err := utils.GetUserIdFromContext(ctx)
+	if err != nil {
+		return err
+	}
+
+	userPerm, err := s.GetUserPermissionCodeNames(ctx, userId)
+	if err != nil {
+		return err
+	}
+
+	if !slices.Contains(userPerm, constant.ManagePermissionGroup) {
+		return ErrForbidCreatePermissionGroup
+	}
+
 	var permissions []Permission
-	
+
 	var checkExistingPermissionsErr error
 	permissions, checkExistingPermissionsErr = s.GetPermissionsByIDs(ctx, permissionGroup.Permissions)
 	if checkExistingPermissionsErr != nil {
@@ -80,24 +107,24 @@ func (s *service) CreatePermissionGroup(ctx context.Context, permissionGroup *Pe
 	// Check for duplicate name
 	existingGroup, err := s.permissionRepo.FindPermissionGroupByName(ctx, permissionGroup.Name)
 	if err == nil && existingGroup != nil {
-		return ErrRoleNameAlreadyExists
+		return ErrPermissionGroupNameAlreadyExists
 	}
 
 	return s.permissionRepo.WithTransaction(ctx, func(txRepo Repository) error {
 		pg := &PermissionGroup{
-			Name:          permissionGroup.Name,
-			Description:   permissionGroup.Description,
-			Permissions:   permissions,
-			CreatedAt:     time.Now(),
+			Name:        permissionGroup.Name,
+			Description: permissionGroup.Description,
+			Permissions: permissions,
+			CreatedAt:   time.Now(),
 		}
 
 		err := txRepo.CreatePermissionGroup(ctx, pg)
 
 		if err != nil {
 			return err
-		} 
-		
-		return nil 
+		}
+
+		return nil
 	})
 }
 
@@ -107,6 +134,20 @@ func (s *service) GetAllPermissionGroups(ctx context.Context, search string, per
 }
 
 func (s *service) GetPermissionGroupByID(ctx context.Context, id uuid.UUID) (*PermissionGroup, error) {
+	userId, err := utils.GetUserIdFromContext(ctx)
+	if err != nil {
+		return nil, err
+	}
+
+	userPerm, err := s.GetUserPermissionCodeNames(ctx, userId)
+	if err != nil {
+		return nil, err
+	}
+
+	if !slices.Contains(userPerm, constant.ViewPermissionGroup) && !slices.Contains(userPerm, constant.ManagePermissionGroup) {
+		return nil, ErrForbidViewPermissionGroup
+	}
+
 	permissionGroup, err := s.permissionRepo.FindPermissionGroupByID(ctx, id)
 	if err != nil {
 		return nil, err
@@ -115,13 +156,27 @@ func (s *service) GetPermissionGroupByID(ctx context.Context, id uuid.UUID) (*Pe
 }
 
 func (s *service) UpdatePermissionGroup(ctx context.Context, id uuid.UUID, permissionGroup *PermissionGroupInput) error {
-	var permissions []Permission
-
-	_, err := s.GetPermissionGroupByID(ctx, id)
+	userId, err := utils.GetUserIdFromContext(ctx)
 	if err != nil {
 		return err
 	}
-	
+
+	userPerm, err := s.GetUserPermissionCodeNames(ctx, userId)
+	if err != nil {
+		return err
+	}
+
+	if !slices.Contains(userPerm, constant.ManagePermissionGroup) {
+		return ErrForbidUpdatePermissionGroup
+	}
+
+	var permissions []Permission
+
+	_, err = s.permissionRepo.FindPermissionGroupByID(ctx, id)
+	if err != nil {
+		return err
+	}
+
 	var checkExistingPermissionsErr error
 	permissions, checkExistingPermissionsErr = s.GetPermissionsByIDs(ctx, permissionGroup.Permissions)
 	if checkExistingPermissionsErr != nil {
@@ -131,9 +186,9 @@ func (s *service) UpdatePermissionGroup(ctx context.Context, id uuid.UUID, permi
 	// Check for duplicate name (excluding current group)
 	duplicateGroup, err := s.permissionRepo.FindPermissionGroupByName(ctx, permissionGroup.Name)
 	if err == nil && duplicateGroup != nil && duplicateGroup.ID != id {
-		return ErrRoleNameAlreadyExists
+		return ErrPermissionGroupNameAlreadyExists
 	}
-	
+
 	pg := &PermissionGroup{
 		Name:        permissionGroup.Name,
 		Description: permissionGroup.Description,
@@ -144,14 +199,28 @@ func (s *service) UpdatePermissionGroup(ctx context.Context, id uuid.UUID, permi
 }
 
 func (s *service) DeletePermissionGroup(ctx context.Context, id uuid.UUID) error {
-	permissionGroup, err := s.GetPermissionGroupByID(ctx, id)
+	userId, err := utils.GetUserIdFromContext(ctx)
 	if err != nil {
 		return err
 	}
-	
+
+	userPerm, err := s.GetUserPermissionCodeNames(ctx, userId)
+	if err != nil {
+		return err
+	}
+
+	if !slices.Contains(userPerm, constant.ManagePermissionGroup) {
+		return ErrForbidDeletePermissionGroup
+	}
+
+	permissionGroup, err := s.permissionRepo.FindPermissionGroupByID(ctx, id)
+	if err != nil {
+		return err
+	}
+
 	if permissionGroup.Name == "Super Admin" {
 		return ErrCannotDeleteSuperAdmin
 	}
-	
+
 	return s.permissionRepo.DeletePermissionGroup(ctx, id)
 }
